Refactor financial balance calculations and UI components
- Removed 'real_balance' from the FinancialBalance model.
- Updated the CalculateFinancialBalanceService to handle recalculations more accurately, including adjustments for total income, expenses, and unidentified amounts.
- Modified SaveFinancialMovementService to ensure expenses and discounts are recorded as negative amounts.
- Created a new migration file for adding indexes to financial tables.
- Enhanced the dashboard view with improved alert handling and chart data processing.
- Simplified the financial balances index view by removing unnecessary conditions.
- Adjusted the edit form for financial balances to use strict equality checks.
- Updated modal summary to correctly format dates and handle wallet data.
- Fixed financial categories index view to calculate totals correctly.
- Improved filter modal for financial movements with better attribute handling.
- Refined financial movements form and modal to ensure proper amount handling.
- Enhanced delete functionality in financial movements table.
- Corrected display of amounts in financial movements table rows.

+++ app/Services/CalculateFinancialBalanceService.php
@@ -34,8 +34,22 @@ class CalculateFinancialBalanceService
             $balance->end_date
         );
 
-        if ($financialMovements->isEmpty() && $balance->initial_balance == 0) {
+        if (
+            $financialMovements->isEmpty() &&
+            $balance->initial_balance == 0 &&
+            $balance->calculated_balance == 0
+        ) {
             return $balance->delete();
+        }
+
+        if (
+            $financialMovements->isEmpty() &&
+            $balance->total_unidentified == 0 &&
+            $balance->total_income == 0 &&
+            $balance->total_expense == 0 &&
+            $balance->calculated_balance != $balance->initial_balance
+        ) {
+            return true;
         }
 
         if ($financialMovements->isEmpty()) {
@@ -62,16 +76,16 @@ class CalculateFinancialBalanceService
                     $balance->total_income += $movement->amount;
                     break;
                 case FinancialMovementType::REFUND:
-                    $balance->total_expense -= $movement->amount;
+                    $balance->total_expense += $movement->amount;
                     break;
                 case FinancialMovementType::DISCOUNT:
-                    $balance->total_income -= $movement->amount;
+                    $balance->total_income += $movement->amount;
                     break;
             }
         }
 
         $balance->total_income = $balance->total_income;
-        $diff = $balance->total_income - $balance->total_expense + $balance->total_unidentified;
+        $diff = $balance->total_income + $balance->total_expense + $balance->total_unidentified;
         $balance->calculated_balance = $balance->initial_balance + $diff;
 
         return $balance->save();
